requirements.txt
asgiref==3.8.1
av==12.3.0
Django==5.0.7
djangorestframework==3.15.2
djitellopy==2.5.0
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.1
